fix async wait issue in loops
async wait is used to auto-wait for element to appear before acting on it. this will hang casperjs/phantomjs if it is done in a loop (eg for loop or while loop). this commit adds code to auto-disable async wait inside the loop and let script continues.

users can intentionally wait by using wait step before the loop, or do some dummy steps that interacts with some element on the webpage, for eg read footer_element into dummy_variable. this will let flow wait for the footer_element to appear before commencing loop