enhance xpath/css selector
old logic is checking for a locator containing given string. for some situations, it does not work for specifying a locator in the shortcut way. for eg there are 2 locators with string user and userid. by providing user, it might map to userid if that element comes before user in the document. the way to do is to use full xpath eg //*[@id="user"]. with this update by providing user, it will maps to user first as it is found to match exactly, even if the elements exists below the userid element.

for continued observation and feedback if this change solves more problems than potentially new problems it introduces. should be a better implementation and give users more accuracy + control of their automation scripts.