improve chrome extension locator
the extension is based on resurrectio project which captures as css selectors and then 'beautifying' into shorter locators. there isn't a perfect implementation as it is a balance between accuracy (no beautifying and simply using the raw css selector) and readability.

to look out for feedback and also from actual field usage to modify accordingly. the chrome extension helps to simplify the user friction in identifying locators instead of using web browser inspector. it is not a replacement for the web browser inspector.

more details of checking identifiers from web browsers can be found in the link in readme.md - https://help.surveygizmo.com/help/how-to-find-element-ids-to-use-with-javascript