start commits towards v2.0
part 2/2 commit to remove the 4 commented off blocks - 

this change is to clean up technical debt from v1.0. 4 blocks of code is used to handle 4 types of contain conditions. now reduced to 1 block and iterate through the 4 types instead. this will make the code easier to maintain. tested ok with test/positive_test and signature.