add code for live mode to tagui_header.js
live mode allows trying out various steps and javascript code
interactively, for steps that are possible in that context

to do so, much of the logic in tagui_parse.php is ported over to
javascript code in tagui_header.js. so that during casperjs execution,
user input can be mapped into interactions with the webpage, and
actioned on the spot to receive feedback.

with this change, it opens up the possibility for future shift to run
through only js without using php/bash/batch. although unlikely,
because initiating through shell/php iteratively handles batch
automation as new separate processes, and helps maintain stability in
automation environment for large repetitive flows.