#37 - navigation of url containing variable
Before this commit, URLs containing variables have to be explicitly set
in { and } block in order to retain the casper context for variables.
With this commit, that is done automatically.

So the lines below will now work. Also, url validation is only done for
raw text url in order not to trigger error for dynamically formed URLs.

link = “abc.com”;
https://'+link+'