#30 - set variable without typing ; at the end
; is used to detect that it is a valid JavaScript code
variable assignments are so often used that detecting = should be
acceptable, since that check is done at the very end of checking for
TagUI steps and other JavaScript keywords.

So the following will now work and be more convenient. Update readme
only when making new release and packaged installations.

a = 1
b = ‘b’
c = ”c”

Adding on, reason for JavaScript checking instead of taking anything
that is not TagUI step as JavaScript is to try to catch as much user
error as possible before even running an automation flow. This speeds
up development time rapidly without going through all the steps during
automation just to run into a user code error.