#32 - baseline option to output to separate folder
not yet tested for windows runner. but that’s ok, shouldn’t break
existing functionality when baseline option is not used. test that when
have access to windows env. tested for macos/linux

decided to put as part of existing framework because it can be done
within the main runners tagui / tagui.cmd easily and elegantly. no need
to wait until extending to natural-language cli helper to implement.