#42 - timeout step to set timeout during execution
previously casper.options.waitTimeout can be used to set timeout
dynamically during execution.

with timeout step it is easier, for example -
timeout 3 seconds
timeout 4.5
timeout 6s

this makes it easy to dynamically adjust before interacting with
certain elements that are expected to take very long to be ready. the
default of 10 seconds can be configured in tagui_config.txt