#42 #26 - improve blocks, loops, live mode
- improve code block and loop handling. to process { and } into casper
blocks if the lines start with { or }. otherwise will cause many valid
javascript code to throw error. such as when using chrome_step or
setting api_config. with this change, the js step is not needed to
avoid those errors. add more error checking for loops to show error at
parsing if { is written on the same line. this will guide users to type
{ on the next line to be auto-parsed into casperjs blocks.

- updated details for steps that are not relevant/supported in live
mode, add support for live mode navigate to url when running chrome
browser. for phantomjs, set window.location.href or using
casper.thenOpen will not work when event loop is not yet passed back to
main 'thread'. also checked that recent changes to steps have been
mirrored to live mode.

- add code in positive test to auto-cleanse local folder path instead
of manually searching and replacing