#53 - bug fix chrome extension tabbing to next field
- update readme on tab key input verbiage (tab keys can’t be captured
as inputs but tabbing to next field works)

- update chrome extension upstream patch from resurrectio on issue when
a form as an element with the name ‘name’ it will have issue

- update chrome extension recorder.js onchange event. the event was
previously changed to enable a custom function in handling drop down
option selection. but there are edge cases such as tabbing to next
field which will not work with that change. this update handle both the
2 scenarios depending on whether the event is for a dropdown option
selection or for everything else (the original resurrection code)