add python 2/3 compability for python integration
this commit allows tagui's python integration to work out of the box whether v2.x or v3.x is installed in the environment. there is no special config needed and works out of the box. note - your python code will still have to be aligned to the python version you are using.