#88 - adding commands and blank lines before URL in flow
TagUI requires the first line of flow script to be an URL, and throws error if that is not the case. This ensures that that user always start the flow file with an URL before any other steps are performed.

However, that means that comments about the flow file can only be entered after the URL first line. This can be improved by having allow comments (lines starting with //) and blank lines before the URL.

This commit facilitates better annotation and makes it easier for users to organize flow files.