#89 - Chrome extension, make it easier to inspect web elements
TagUI Chrome extension has a right-click shortcut menu to note down element. A comment will be in the recorded script giving the element identifier.

However, in some cases, users want to get the identifiers immediately as they browse through different parts of the website, for example, when writing the automation flow file manually. This upcoming commit will add an 'inspect element' on the right-click menu, to popup and show the identifier immediately while recording.