#97 - mute 'about:blank - '
The 3rd and final commit for this feature. 'about:blank - ' will be muted from output, eg for desktop automation workflows where no website access to web app is needed.

as the changes for this feature have large technical impact to code, made the commits separately for better tracking and future reference.