#106 - rollback as it can break result of existing scripts
for example with the change in previous commit, an echo step that happens before a dump step will end up executing and showing the output after dump step. because echo command is entered into the event queue but not dump. to look for more data points. if there is a change, it might be required in large number of steps, such as - download, receive, dump, write, load, check, api, dom, js, r, py, vision