support datatables in test mode
details of issue here - https://github.com/kelaberetiv/TagUI/issues/348

when datatable is used with test option, execution hangs after the first iteration. this is because casperjs test mode treats every parameter as in input file to test. when the datatable parameter is passed to casperjs it hangs because it is not a valid test js file. modified tagui and tagui.cmd to not pass parameters to casperjs in test mode.

technical details here - http://docs.casperjs.org/en/latest/testing.html#arguments

`The casperjs test command will treat every passed argument as file or directory paths containing tests. It will recursively scan any passed directory to search for *.js or *.coffee files and add them to the stack.`