#372 - Updated language definitions and translation engine for new steps
Creating and sending a PR for updated TagUI language definitions and translation engine for the new steps (`keyboard`, `mouse`) and helper functions (`mouse_xy()`, `mouse_x()`, `mouse_y()`) created in #370, #368, #361 and #365.

This lets users use the new steps and helper functions in 21 different languages. More details on the multiple languages feature of TagUI in below link of documentation -

https://github.com/kelaberetiv/TagUI#native-languages