#347 - live mode dynamic variable patch (#379)
When \\ is used, it becomes executed as \. When \t is used it becomes executed as tab. This is because eval() is used to derive dynamic variables.

This patch converts \ in input to \\ so that the input gets executed as intended and retains its original form after passing through the eval function.