#384 - run results auditing and tracking enhancement to report option
Sending a PR to enhance report option. In current state when report option is used, TagUI will generate a HTML log file, in addition to the usual text log file.

With this enhancement, when the report option is used, TagUI will track the run results in a csv spreadsheet (tagui/src/tagui_report.csv), along with html logs that persist with a running number.

This allows users to explicitly decide which automation flows to be tracked and tabulated in tagui_report.csv. This information may be useful for personal reference, to measure time savings from automation, for keeping a trail for auditing purpose, or perhaps facilitate sending a summary to boss. Following is an example of how tagui_report.csv will look like (time taken in seconds) -

#| AUTOMATION FLOW | START TIME | FINISH TIME | ERROR STATUS | LOG FILE
:--|:--|:--|:--|:--|:--
1 | /Users/kensoh/Desktop/download_flow | Sun Apr 07 2019 17:33:58 GMT+0800 (+08) | 32.1 | SUCCESS | /Users/kensoh/Desktop/download_flow_1.html
2 | /Users/kensoh/Desktop/upload_flow | NOT STARTED | NOT FINISHED | [LINE 1] cannot understand step self-destruct | /Users/kensoh/Desktop/upload_flow_2.html
3 | /Users/kensoh/Desktop/update_flow | NOT STARTED | NOT FINISHED | [LINE 1] cannot understand step reboot computer | /Users/kensoh/Desktop/update_flow_3.html
4 | /Users/kensoh/Desktop/booking_flow | Sun Apr 07 2019 17:39:00 GMT+0800 (+08) | NOT FINISHED | cannot find confirm_booking | /Users/kensoh/Desktop/booking_flow_4.html