#423 - enable check step within for loop
https://github.com/kelaberetiv/TagUI/issues/423

update summary - check step was not originally designed to work in for loops. this update allows check step to work in for loops, and also adds test cases to verify that check step can work within for loops.

technical details - check step will insert an if block into the code block that immediately closes without normal processing line by line. this will cause TagUI to be mistaken that parsing is still within an if block and not handling a for loop block as it should be handled. in this change, a helper function check_intent_clear_injected_if_block() is added and called during parsing of check step to update TagUI internal tracking that the if block has been closed at end of the check step. that allows for loops to function correctly when check step is used within a loop.