#443 - timeout step sikulix variable update (#453)
SikuliX wait_timeout also needs to be changed as part of sikuli_step() when timeout step is used. Otherwise, when present() or visible() is invoked, the SikuliX timeout setting gets reverted back to its internal variable.

Also, 'casper' object should be used in displaying the error message instead of 'this'. Otherwise, for the execution context, this would be undefined and error message does not show up.