#461 - support defining regions for OCR and snapshot (#463)
Following up on #457, where transparency (0% opacity) in images is supported in visual automation for advance use cases, raising a PR that allows defining regions for steps read, show, save, snap. This allows users to define a square/rectangular region on the screen to perform OCR for text extraction, and saving screenshot of that region.